refactor lofire-broker & lofire-p2p types
