+++ lofire-broker/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 
 [dependencies]
 lofire = { path = "../lofire" }
+lofire-p2p = { path = "../lofire-p2p" }
 blake3 = "1.3.1"
 chacha20 = "0.9.0"
 serde = { version = "1.0", features = ["derive"] }
